refactor(ui): replace ui_has(kUIMultigrid) with multigrid specific events
Problem:
The code has a lot of ui_has(kUIMultigrid) checks, which can cause
different behaviour when a multigrid is connected to the same instance
as a single grid UI, and therefore affect it as well.

And even when the behaviour is stays the same, the events will be
wrongly sent to the single grid UI as well, even if it's not supposed to
receive multigrid events.

Solution:
Tag the multigrid UI events to be sent to only multigrid UIs and remove
checks around the codebase.